ren.c:vi.c: silence compiler

+++ ren.c
@@ -101,7 +101,7 @@ int *ren_position(char *s, char ***chrs, int *n)
 	pos = emalloc(((nn + 1) * sizeof(pos[0])) * 2);
 	if (xorder && dir_reorder(chrs[0], pos, nn)) {
 		off = &pos[nn+1];
-		memcpy(off, pos, nn * sizeof(pos[0]));
+		memcpy(off, pos, (int)(nn * sizeof(pos[0])));
 		for (i = 0; i < nn; i++) {
 			pos[off[i]] = cpos;
 			cpos += ren_cwid(chrs[0][off[i]], cpos);
